Rename core/log_timings.py to core/timings.py
Splitting the timing hooks left two files named log_timings.py, one in core
and one in implementations/hooks, distinguished only by their class names.

Renaming the core half resolves it in the direction that matches both
packages' conventions: core names modules after the concept (hooks.py,
errors.py, problem.py), while implementations/hooks uses a log_ prefix
(log_errors.py, log_work.py, log_solution.py). GPUTimings therefore keeps
the log_timings.py path it has always had, so no public import changes.
