+++ pySDC/core/controller.py
@@ -8,7 +8,7 @@ from pySDC.core.base_transfer import BaseTransfer
 from pySDC.helpers.pysdc_helper import FrozenClass
 from pySDC.core.check_convergence import CheckConvergence
 from pySDC.core.default_hook import DefaultHooks
-from pySDC.core.log_timings import CPUTimings
+from pySDC.core.timings import CPUTimings
 
 
 # short helper class to add params as attributes
